# backend/application/src/main/resources/api/v1/specs/openapi.yaml
openapi: 3.0.3
info:
  title: Strategy & Planning API
  version: 0.1.0
  description: |
    Strategy & Planning toolkit API. Provides endpoints for the Case Study Builder,
    Category Analysis Builder, and RFP Outline Generator tools. Authentication via Clerk SSO (Bearer JWT).
servers:
  - url: /api/v1
    description: V1 base path (DOCUMENTATION ONLY — the openapi-generator-maven-plugin
      does NOT prepend this to generated Spring routes. Path keys below carry the
      `/api/v1/` prefix verbatim. See canonical-openapi-rules.md → "/api/v1 prefix").
security:
  - bearerAuth: [ ]
...
paths:
...
  /api/v1/category-analyses:
    get:
      tags: [ CategoryAnalyses ]
      operationId: listCategoryAnalyses
      summary: List all category analyses for the authenticated user
      description: Returns all category analyses created by the authenticated user, ordered by creation date descending.
      security: [ { bearerAuth: [ ] } ]
      responses:
        "200":
          description: List of category analysis summaries for the authenticated user.
          content:
            application/json:
              schema:
                type: array
                description: Ordered list of category analysis summaries.
                items:
                  $ref: "#/components/schemas/CategoryAnalysisSummaryV1"
        "401":
          description: The authorization token not provided or expired.
        "403":
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "500":
          description: Exception occurred while listing category analyses.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"

    post:
      tags: [ CategoryAnalyses ]
      operationId: createCategoryAnalysis
      summary: Create and submit a new category analysis
      description: |
        Creates a new category analysis with the provided form fields.
        The record is persisted with status SUBMITTED, indicating it is ready for Google Slides
        generation once OAuth is configured. Outputs a 3-5 slide deck narrative.
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        required: true
        description: Category analysis form fields. Title is required; all other fields are optional.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryAnalysisRequestV1"
      responses:
        "201":
          description: Category analysis created and submitted successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryAnalysisV1"
        "400":
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppValidationExceptionResponseV1"
        "401":
          description: The authorization token not provided or expired.
        "403":
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "500":
          description: Exception occurred while creating the category analysis.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
  /api/v1/category-analyses/{id}:
    get:
      tags: [ CategoryAnalyses ]
      operationId: getCategoryAnalysis
      summary: Get a category analysis by ID
      description: Returns the full detail of a category analysis belonging to the authenticated user.
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the category analysis.
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Category analysis detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryAnalysisV1"
        "401":
          description: The authorization token not provided or expired.
        "403":
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "404":
          description: Category analysis not found or does not belong to the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "500":
          description: Exception occurred while retrieving the category analysis.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
  /api/v1/category-analyses/standard-connections:
    get:
      tags: [ CategoryAnalyses ]
      operationId: getStandardConnections
      summary: Report connection status for the Standard Category Analysis generator
      description: |
        Reports whether the AI drafting engine is configured (OpenAI API key present) and whether
        the authenticated user's Google account grants Drive/Slides access via Clerk SSO.
      security: [ { bearerAuth: [ ] } ]
      responses:
        "200":
          description: Connection status for the standard deck generator.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardConnectionsV1"
        "401":
          description: The authorization token not provided or expired.
        "403":
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "500":
          description: Exception occurred while checking connections.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
  /api/v1/category-analyses/standard-drafts:
    post:
      tags: [ CategoryAnalyses ]
      operationId: draftStandardCategoryAnalysis
      summary: AI-draft all fields of a Standard Category Analysis
      description: |
        Uses the AI drafting engine to produce every text field of the Standard Category Analysis
        template (5 slides, ~50 fields) from a short brief. The result is
        meant to be reviewed and edited by the user before deck creation. Stats and sources are
        AI-drafted and must be verified before client use.
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        required: true
        description: Short brief for the analysis. Category and client name are required.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StandardDraftRequestV1"
      responses:
        "200":
          description: Complete drafted field set for review.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardDraftV1"
        "400":
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppValidationExceptionResponseV1"
        "401":
          description: The authorization token not provided or expired.
        "403":
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "502":
          description: AI drafting engine not configured or the drafting call failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "500":
          description: Exception occurred while drafting the analysis.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
  /api/v1/category-analyses/standard-drafts/slide:
    post:
      tags: [ CategoryAnalyses ]
      operationId: redraftStandardCategorySlide
      summary: AI-redraft the fields of a single slide of a Standard Category Analysis
      description: |
        Re-drafts only the text fields of one slide of the
        Standard Category Analysis template, using the original brief plus an optional per-slide
        guidance note. The current values of the whole deck are supplied as context so the redraft
        stays consistent with the slides the user is keeping. Only the returned slide's fields
        should replace existing values; edits on other slides are preserved by the caller. Stats
        and sources are AI-drafted and must be verified before client use.
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        required: true
        description: Brief, slide number, optional per-slide note, and current deck values for context.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedraftSlideRequestV1"
      responses:
        "200":
          description: Redrafted field set for the requested slide.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardDraftV1"
        "400":
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppValidationExceptionResponseV1"
        "401":
          description: The authorization token not provided or expired.
        "403":
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "502":
          description: AI drafting engine not configured or the drafting call failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "500":
          description: Exception occurred while redrafting the slide.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
  /api/v1/category-analyses/standard-decks:
    post:
      tags: [ CategoryAnalyses ]
      operationId: createStandardCategoryDeck
      summary: Create the Standard Category Analysis deck in the user's Google Slides
      description: |
        Copies the master Standard Category Analysis template into the authenticated user's own
        Google Drive, fills every template token with the reviewed field values, replaces the
        image placeholders on slides 1-4 with friendly, on-vertical generated photographs, and
        returns the new presentation's URL. Requires the user's Google account (via Clerk SSO) to
        grant Drive and Slides access.
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        required: true
        description: Reviewed field values for deck creation.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStandardDeckRequestV1"
      responses:
        "201":
          description: Deck created in the user's Google Drive.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryAnalysisV1"
        "400":
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppValidationExceptionResponseV1"
        "401":
          description: The authorization token not provided or expired.
        "403":
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "502":
          description: Google Drive/Slides access not granted or the Google API call failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "500":
          description: Exception occurred while creating the deck.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    GenerationStatusV1:
      type: string
      description: |
        Lifecycle status of a generated document.
        DRAFT — saved but not yet submitted for generation.
        SUBMITTED — submitted for Google Slides generation.
        GENERATED — Google Slides deck created in the user's Drive.
      enum: [ DRAFT, SUBMITTED, GENERATED ]
    GoogleConnectionStatusV1:
      type: string
      description: |
        State of the caller's Google connection for Drive/Slides work.
        NOT_CONNECTED — the user has not signed in with Google.
        MISSING_SCOPES — signed in with Google, but the Drive/Slides scopes were not granted.
        CONNECTED — Drive & Slides access granted.
      enum: [ NOT_CONNECTED, MISSING_SCOPES, CONNECTED ]
...
    CategoryAnalysisSummaryV1:
      type: object
      additionalProperties: false
      description: Lightweight summary of a category analysis for list views.
      required: [ id, title, status, createdAt ]
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the category analysis.
        title:
          type: string
          description: Title of the category analysis.
        category:
          type: string
          nullable: true
          description: Market category or sector being analysed.
        slidesUrl:
          type: string
          nullable: true
          description: URL of the generated Google Slides presentation in the user's Drive, when created.
        status:
          description: Lifecycle state of the category analysis deck generation.
          allOf:
            - $ref: "#/components/schemas/GenerationStatusV1"
        createdAt:
          type: string
          format: date-time
          description: UTC timestamp when the category analysis was created.
    CategoryAnalysisV1:
      type: object
      additionalProperties: false
      description: Full detail of a category analysis. Outputs a 3-5 slide deck narrative.
      required: [ id, title, status, createdAt ]
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the category analysis.
        title:
          type: string
          description: Title of the category analysis.
        category:
          type: string
          nullable: true
          description: Market category or sector being analysed.
        marketOverview:
          type: string
          nullable: true
          description: High-level overview of the market landscape.
        keyPlayers:
          type: string
          nullable: true
          description: Key competitors or players in the category.
        trends:
          type: string
          nullable: true
          description: Emerging trends shaping the category.
        opportunities:
          type: string
          nullable: true
          description: Strategic opportunities identified within the category.
        recommendations:
          type: string
          nullable: true
          description: Recommended actions or strategies.
        sourceFileNames:
          type: string
          nullable: true
          description: Comma-separated names of uploaded source documents used for analysis.
        slidesUrl:
          type: string
          nullable: true
          description: URL of the generated Google Slides presentation in the user's Drive, when created.
        templateKind:
          type: string
          nullable: true
          description: Template used for generation (e.g. "STANDARD"), null for manual submissions.
        status:
          description: Lifecycle state of the category analysis deck generation.
          allOf:
            - $ref: "#/components/schemas/GenerationStatusV1"
        createdAt:
          type: string
          format: date-time
          description: UTC timestamp when the category analysis was created.
    CreateCategoryAnalysisRequestV1:
      type: object
      additionalProperties: false
      description: Request payload for creating a new category analysis. Title is required; all other fields are optional.
      required: [ title ]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Title of the category analysis.
        category:
          type: string
          maxLength: 255
          nullable: true
          description: Market category or sector being analysed.
        marketOverview:
          type: string
          maxLength: 5000
          nullable: true
          description: High-level overview of the market landscape.
        keyPlayers:
          type: string
          maxLength: 5000
          nullable: true
          description: Key competitors or players in the category.
        trends:
          type: string
          maxLength: 5000
          nullable: true
          description: Emerging trends shaping the category.
        opportunities:
          type: string
          maxLength: 5000
          nullable: true
          description: Strategic opportunities identified within the category.
        recommendations:
          type: string
          maxLength: 5000
          nullable: true
          description: Recommended actions or strategies.
        sourceFileNames:
          type: string
          maxLength: 2000
          nullable: true
          description: Comma-separated names of uploaded source documents used for analysis.
    StandardConnectionsV1:
      type: object
      additionalProperties: false
      description: Connection status for the Standard Category Analysis deck generator.
      required: [ aiConnected, googleConnected, googleStatus ]
      properties:
        aiConnected:
          type: boolean
          description: True when the AI drafting engine (OpenAI API key) is configured.
        googleConnected:
          type: boolean
          description: True when the authenticated user's Google account grants Drive/Slides access via Clerk SSO.
        googleStatus:
          description: Detailed Google connection state, so the UI can guide the user through re-consent.
          allOf:
            - $ref: "#/components/schemas/GoogleConnectionStatusV1"
    StandardDraftRequestV1:
      type: object
      additionalProperties: false
      description: Short brief used to AI-draft a Standard Category Analysis.
      required: [ category, clientName, clientWebsite ]
      properties:
        category:
          type: string
          maxLength: 255
          minLength: 1
          description: Market category or industry being analysed (e.g. "Regional Travel", "Western Boots").
        clientName:
          type: string
          maxLength: 255
          minLength: 1
          description: Client the analysis is prepared for (e.g. "Alberta Boot").
        clientWebsite:
          type: string
          maxLength: 2048
          minLength: 1
          description: >-
            Client's website URL (e.g. "https://albertaboot.com"). The builder reviews the
            site to understand the client's business focus and confirm the deck aligns with it.
        storyTheme:
          type: string
          maxLength: 500
          nullable: true
          description: >-
            Optional narrative theme to guide the arc of the deck (e.g. "how adopting CTV can
            increase market awareness"). Emphasised across slides while still covering broader trends.
        guidanceNotes:
          type: string
          maxLength: 5000
          nullable: true
          description: Optional focus areas, angles, or constraints to steer the draft.
    StandardDraftFieldV1:
      type: object
      additionalProperties: false
      description: One drafted template field with its slide grouping metadata.
      required: [ key, label, slideNumber, value ]
      properties:
        key:
          type: string
          description: Stable template token key (e.g. "trends_headline").
        label:
          type: string
          description: Human-readable label shown in the review UI.
        slideNumber:
          type: integer
          format: int32
          description: Slide the field belongs to (1-5).
        value:
          type: string
          description: Drafted text value; editable by the user before deck creation.
    DraftAlignmentV1:
      type: object
      additionalProperties: false
      description: Website-based confirmation that the drafted deck matches the client's business focus.
      required: [ confirmed, matches, message ]
      properties:
        clientBusinessFocus:
          type: string
          nullable: true
          description: Short summary of the client's core business focus, derived from their website.
        confirmed:
          type: boolean
          description: >-
            Whether the client website could be read to actually check alignment. When false the
            match result is unverified and the message explains why.
        matches:
          type: boolean
          description: >-
            Whether the analysed category clearly fits the client's business focus. Only meaningful
            when confirmed is true.
        message:
          type: string
          description: Human-readable confirmation or warning about topic/business alignment.
    StandardDraftV1:
      type: object
      additionalProperties: false
      description: Complete AI-drafted field set for the Standard Category Analysis template.
      required: [ fields, alignment ]
      properties:
        fields:
          type: array
          description: All drafted template fields ordered by slide.
          items:
            $ref: "#/components/schemas/StandardDraftFieldV1"
        alignment:
          description: How closely the draft matched the requested category and client.
          allOf:
            - $ref: "#/components/schemas/DraftAlignmentV1"
    RedraftSlideRequestV1:
      type: object
      additionalProperties: false
      description: Brief plus current deck context used to AI-redraft a single slide of a Standard Category Analysis.
      required: [ category, clientName, slideNumber, currentFields ]
      properties:
        category:
          type: string
          maxLength: 255
          minLength: 1
          description: Market category or industry being analysed (e.g. "Regional Travel", "Western Boots").
        clientName:
          type: string
          maxLength: 255
          minLength: 1
          description: Client the analysis is prepared for (e.g. "Alberta Boot").
        guidanceNotes:
          type: string
          maxLength: 5000
          nullable: true
          description: Original brief focus areas or constraints, may be null.
        slideNumber:
          type: integer
          format: int32
          minimum: 1
          maximum: 5
          description: Slide to redraft (1-5).
        slideNote:
          type: string
          maxLength: 5000
          nullable: true
          description: Optional guidance for this redraft (e.g. "make it more data-driven", "focus on Gen Z").
        currentFields:
          type: array
          maxItems: 200
          description: Current values for every template field of the deck, supplied as context so the redraft stays consistent with slides being kept.
          items:
            $ref: "#/components/schemas/StandardFieldValueV1"
    StandardFieldValueV1:
      type: object
      additionalProperties: false
      description: A reviewed field value keyed by its template token.
      required: [ key, value ]
      properties:
        key:
          type: string
          minLength: 1
          maxLength: 255
          description: Stable template token key.
        value:
          type: string
          maxLength: 20000
          description: Final reviewed text value.
    CreateStandardDeckRequestV1:
      type: object
      additionalProperties: false
      description: Reviewed field values used to create the deck in the user's Google Slides.
      required: [ category, clientName, fields ]
      properties:
        category:
          type: string
          maxLength: 255
          minLength: 1
          description: Market category or industry being analysed.
        clientName:
          type: string
          maxLength: 255
          minLength: 1
          description: Client the analysis is prepared for.
        fields:
          type: array
          maxItems: 200
          description: Final reviewed values for every template field.
          items:
            $ref: "#/components/schemas/StandardFieldValueV1"
...
    AppValidationExceptionResponseV1:
      type: object
      additionalProperties: false
      description: Representation of validation exception occurred in REST API format.
      required: [ errors, timestamp, correlationId ]
      properties:
        errors:
          description: The list of validation errors occurred.
          type: array
          items:
            $ref: "#/components/schemas/FieldToErrorResponseV1"
        timestamp:
          type: string
          format: date-time
          description: The timestamp exception occurred.
        correlationId:
          type: string
          description: The log stream UUID.
    FieldToErrorResponseV1:
      type: object
      additionalProperties: false
      description: Representation of validation error bound to a field or semantic request part.
      required: [ code, field, error ]
      properties:
        code: { type: string, description: "The code of validation error occurred." }
        field: { type: string, description: "The field or semantic request part not passed validation." }
        error: { type: string, description: "The validation error occurred." }
    AppApiExceptionResponseV1:
      type: object
      additionalProperties: false
      description: Representation of exception occurred in REST API format.
      required: [ code, message, timestamp, correlationId ]
      properties:
        code: { type: string, description: "The code of exception occurred." }
        message: { type: string, description: "The message of exception occurred." }
        timestamp: { type: string, format: date-time, description: "The timestamp exception occurred." }
        correlationId: { type: string, description: "The log stream UUID." }